WIP: Fix install script. Enable passing google cloud key.

+++ install.sh
@@ -1,10 +1,33 @@
+#!/bin/bash
+error() {
+  printf '\E[31m'; echo "$@"; printf '\E[0m'
+}
+
+#if [[ $EUID -eq 0 ]]; then
+if [ "$(id -u)" != "0" ]; then
+    error "This script should be run using sudo or as the root user"
+    exit 1
+fi
+
+if [ "$1" == "" ]; then
+    error "Please provide google cloud access key"
+    exit 1
+fi
+
+if [ ! -f "$1" ]; then
+    error "Key not found!"
+    exit 1
+fi
+
+GS_KEY=$1
+
 # install mongodb
-sudo apt-get install -y libssl-dev python-pip supervisor wget mongodb-org openssl
+apt-get install -y libssl-dev python-pip supervisor wget mongodb-server openssl curl
 
-sudo service supervisor restart
+service supervisor restart
 
-sudo service mongod enable
-sudo service mongod start
+#service mongodb enable
+service mongodb start
 
 # install gsutil
 pip install pyopenssl
@@ -12,40 +35,44 @@ pip install pyopenssl
 export CLOUD_SDK_REPO="cloud-sdk-$(lsb_release -c -s)"
 
 # Add the Cloud SDK distribution URI as a package source
-echo "deb https://packages.cloud.google.com/apt $CLOUD_SDK_REPO main" | sudo tee -a /etc/apt/sources.list.d/google-cloud-sdk.list
+rm /etc/apt/sources.list.d/google-cloud-sdk.list || true
+
+echo "deb https://packages.cloud.google.com/apt $CLOUD_SDK_REPO main" | tee -a /etc/apt/sources.list.d/google-cloud-sdk.list
 
 # Import the Google Cloud Platform public key
-curl https://packages.cloud.google.com/apt/doc/apt-key.gpg | sudo apt-key add -
+curl https://packages.cloud.google.com/apt/doc/apt-key.gpg | apt-key add -
 
 # Update the package list and install the Cloud SDK
-sudo apt-get update && sudo apt-get install -y google-cloud-sdk
+apt-get update && apt-get install -y google-cloud-sdk
 
 # fetch the proxy program (binary file)
-gcloud auth activate-service-account --key-file gc_credentials.json
+gcloud auth activate-service-account --key-file ${GS_KEY}
 
 gsutil -m cp gs://pos-proxy/staging/bin/pos-proxy .
+pwd
 
-cp pox_proxy /usr/local/bin/pos_proxy
+mkdir -p /usr/local/bin
+cp ./pos-proxy /usr/local/bin/pos-proxy
 
-sudo chmod +x /usr/local/bin/pos_proxy
+chmod +x /usr/local/bin/pos-proxy
 
 # create supervisor configuration file
-sudo touch /etc/supervisor/conf.d/pos_proxy.conf
-read -r -d '' CONF <<- EOM
-[program:long_script]
-command=/usr/local/bin/pos_proxy.sh
+FILE=/etc/supervisor/conf.d/pos_proxy.conf
+touch $FILE
+cat <<EOM >$FILE
+[program:pos-proxy]
+command=/usr/local/bin/pos-proxy
 autostart=true
 autorestart=true
 stderr_logfile=/var/log/pos_proxy.err.log
 stdout_logfile=/var/log/pos_proxy.out.log
 EOM
 
-echo "$CONF" > /etc/supervisor/conf.d/pos_proxy.conf
-
 # copy the configuration file to/etc/cloudinn/pos_config.json
-sudo mkdir /etc/cloudinn
-sudo touch /etc/cloudinn/pos_config.json
-read -r -d '' pos_config <<- EOM
+mkdir -p /etc/cloudinn || true
+FILE=/etc/cloudinn/pos_config.json
+touch $FILE
+cat <<EOM >$FILE
 {
 	"backend_uri": "https://staging.cloudinn.net",
     "tenant_id": 4,
@@ -63,4 +90,3 @@ read -r -d '' pos_config <<- EOM
      ]
 }
 EOM
-sudo echo "$pos_config" > /etc/cloudinn/pos_config.json
